Extract Cardinality to separate file
Incidentally, Cardinality was defined twice so this commit removes that
duplication.
